Add simple tests and implementation figures

+++ FiguresSolution.Figures/Interfaces/ICalculateArea.cs
@@ -8,6 +8,6 @@ namespace FiguresSolution.Figures.Interfaces
 {
     public interface ICalculateArea
     {
-        int Area { get; }
+        double Area { get; }
     }
 }
